Fix payroll generated restriction

+++ app/Http/Controllers/DeductionController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Actions\Hr\IsAccountingPeriodLocked;
 use App\Models\Deduction;
 use App\Models\Employee;
 use App\Models\Payroll;
@@ -51,20 +52,12 @@ class DeductionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(StoreDeductionRequest $request)
-    {
-        //
-        $accounting_system_id = $this->request->session()->get('accounting_system_id');
-        // Disable create if payroll is generated for current accounting period.
-        $payroll = Payroll::where('accounting_system_id', $accounting_system_id)->get();
-        if(!$payroll->isEmpty())
-        return redirect()->back()->with('danger', 'Payroll already created! Unable to generate new deduction in current accounting period.');
-            
-        
+    {            
         for($i = 0; $i < count($request->employee); $i++)
         {        
             // Store
                $deduction = new Deduction;
-               $deduction->accounting_system_id = $accounting_system_id;
+               $deduction->accounting_system_id = session('accounting_system_id');
                $deduction->employee_id = $request->employee[$i]->value;
                $deduction->date = $request->date;
                $deduction->price = $request->price[$i];
@@ -105,13 +98,11 @@ class DeductionController extends Controller
      */
     public function update(Request $request, Deduction $deduction)
     {
-        //
-        $accounting_system_id = $this->request->session()->get('accounting_system_id');
         // Disable create if payroll is generated for current accounting period.
-        $payroll = Payroll::where('accounting_system_id', $accounting_system_id)->get();
-        if(!$payroll->isEmpty())
-        return redirect()->back()->with('danger', 'Payroll already created! Unable to update deduction in current accounting period.');
-          
+        if(IsAccountingPeriodLocked::run($deduction->date))
+            return redirect()->back()->with('danger', 'Payroll already created! Unable to update deduction in current accounting period.');
+        
+        // TODO: Update Deduction
     }
 
     /**
@@ -122,21 +113,18 @@ class DeductionController extends Controller
      */
     public function destroy(Deduction $deduction)
     {
-          //
-          $accounting_system_id = $this->request->session()->get('accounting_system_id');
-          // Disable create if payroll is generated for current accounting period.
-          $payroll = Payroll::where('accounting_system_id', $accounting_system_id)->get();
-          if(!$payroll->isEmpty())
-          return redirect()->back()->with('danger', 'Payroll already created! Unable to delete deduction in current accounting period.');
-            
-          if(isset($deduction->payroll_id))
-          {
-              return redirect()->back()->with('danger', 'Deduction already pending in payroll.');
-          }
-          else
-          {
-              $deduction->delete();
-              return redirect()->back()->with('success', 'Deduction has been deleted.');
-          }           
+        // Disable create if payroll is generated for current accounting period.
+        if(IsAccountingPeriodLocked::run($deduction->date))
+            return redirect()->back()->with('danger', 'Payroll already created! Unable to delete deduction in current accounting period.');
+        
+        if(isset($deduction->payroll_id))
+        {
+            return redirect()->back()->with('danger', 'Deduction already pending in payroll.');
+        }
+        else
+        {
+            $deduction->delete();
+            return redirect()->back()->with('success', 'Deduction has been deleted.');
+        }           
     }
 }
