+++ app/Http/Controllers/EmployeeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Actions\Hr\IsAccountingPeriodLocked;
 use App\Models\Employee;
 use App\Models\Payroll;
 use App\Models\Settings\ChartOfAccounts\AccountingPeriods;
@@ -28,9 +29,8 @@ class EmployeeController extends Controller
         $accounting_system_id = $this->request->session()->get('accounting_system_id');
         
         // Disable creation of new employees if payroll is generated for current accounting period.
-        $payroll = Payroll::where('accounting_system_id', $accounting_system_id)->get();
-        if(!$payroll->isEmpty())
-        return back()->with('error', 'Payroll already created! Unable to generate new employee in current accounting period.');
+        if(IsAccountingPeriodLocked::run($request->date_started_working))
+            return back()->with('error', 'Payroll already created! Unable to generate new employee in current accounting period.');
         
         // Create Employee
         Employee::create([
@@ -60,14 +60,8 @@ class EmployeeController extends Controller
         // TODO: Restrictions if any
         $accounting_system_id = $this->request->session()->get('accounting_system_id');
         // Get range of current accounting period
-        $min = AccountingPeriods::where('accounting_system_id', $accounting_system_id)
-        ->min('date_from');
-        $max = AccountingPeriods::where('accounting_system_id', $accounting_system_id)
-        ->max('date_to');
-
-        // Disable edit if date hired is not within range
-        if($request->date_started_working < $min || $request->date_started_working > $max)
-        return back()->with('error', 'Error Updating! Start of working date should be within accounting period range.');
+        if(IsAccountingPeriodLocked::run($request->date_started_working))
+            return back()->with('error', 'Start of working date should be within accounting period range.');
 
         $employee->update([
             'accounting_system_id' => $accounting_system_id,
@@ -91,7 +85,8 @@ class EmployeeController extends Controller
 
     public function destroy(Employee $employee)
     {
-        // TODO: Restrictions if any
+        if(IsAccountingPeriodLocked::run($employee->date_started_working))
+            return back()->with('error', "Can't delete employee. Employee is already included in payroll.");
 
         $employee->delete();
         return back()->with('success', 'Successfully deleted an employee record.');
